+++ examples/demo-rollup/tests/wallet/mod.rs
@@ -97,7 +97,7 @@ fn test_display_unsigned_tx() {
                 &unsigned_data
             )
             .unwrap(),
-        r#"{ runtime_call: Bank.Mint { coins: 10000 coins of token ID token_1zut3w9chzut3w9chzut3w9chzut3w9chzut3w9chzut3w9chzutsuzalks, mint_to_address: sov1pv9skzctpv9skzctpv9skzctpv9skzctpv9skzctpv9skqm7ehv }, generation: 0, details: { max_priority_fee_bips: 0, max_fee: 1000000000, gas_limit: [10000000, 10000000], chain_id: 4321 } }"#
+        r#"{ runtime_call: Bank.Mint { coins: 10000 coins of token ID token_1zut3w9chzut3w9chzut3w9chzut3w9chzut3w9chzut3w9chzutsuzalks, mint_to_address: sov1pv9skzctpv9skzctpv9skzctpv9skzctpv9skzctpv9skqm7ehv }, generation: 0, details: { max_priority_fee_bips: 0, max_fee: 100000000000, gas_limit: [1000000000, 1000000000], chain_id: 4321 } }"#
     );
 }
 
@@ -127,6 +127,6 @@ fn test_display_signed_tx() {
                 &signed_data
             )
             .unwrap(),
-        format!("{{ signature: {{ msg_sig: 0x{} }}, pub_key: {{ pub_key: 0x{} }}, runtime_call: Bank.Mint {{ coins: 10000 coins of token ID token_1zut3w9chzut3w9chzut3w9chzut3w9chzut3w9chzut3w9chzutsuzalks, mint_to_address: sov1pv9skzctpv9skzctpv9skzctpv9skzctpv9skzctpv9skqm7ehv }}, generation: 0, details: {{ max_priority_fee_bips: 0, max_fee: 1000000000, gas_limit: [10000000, 10000000], chain_id: 4321 }} }}", signature_display, pubkey_display)
+        format!("{{ signature: {{ msg_sig: 0x{} }}, pub_key: {{ pub_key: 0x{} }}, runtime_call: Bank.Mint {{ coins: 10000 coins of token ID token_1zut3w9chzut3w9chzut3w9chzut3w9chzut3w9chzut3w9chzutsuzalks, mint_to_address: sov1pv9skzctpv9skzctpv9skzctpv9skzctpv9skzctpv9skqm7ehv }}, generation: 0, details: {{ max_priority_fee_bips: 0, max_fee: 100000000000, gas_limit: [1000000000, 1000000000], chain_id: 4321 }} }}", signature_display, pubkey_display)
     );
 }